+++ tests/create_suite.py
@@ -10,10 +10,9 @@ DATADIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data")
 VCFS = ["bcftools.vcf", "multisample.vcf", "simulated.vcf"]
 
 
-def assert_correct(vcf_online: str, vcf_offline: str):
-    """
-    Check that two VCF files annotated using online and offline mode are
-    identical.
+def assert_correct_vcf(vcf_online: str, vcf_offline: str):
+    """Check that two VCF files annotated using online and offline mode
+    are identical.
 
     :param str vcf_online: vcf file annotated with online mode
 
@@ -26,9 +25,23 @@ def assert_correct(vcf_online: str, vcf_offline: str):
         assert vcf1.read() == vcf2.read()
 
 
-def check_annotations(input_vcf: str):
+def assert_correct_csv(csv_online: str, csv_offline: str):
+    """Check that two CSV files annotated using online and offline mode
+    are identical.
+
+    :param str csv_online: csv file annotated with online mode
+
+    :param str csv_offline: csv file annotated with offline mode
+
+    :return:
     """
-    Perform assert_online_offline() on the set of online and offline
+    assert os.path.getsize(csv_online) == os.path.getsize(csv_offline)
+    with open(csv_online) as csv1, open(csv_offline) as csv2:
+        assert csv1.read() == csv2.read()
+
+
+def check_annotations(input_vcf: str):
+    """Perform assert_online_offline() on the set of online and offline
     annotated VCF files from the starting VCF file.
 
     :param str input_vcf: input vcf file name
@@ -38,32 +51,32 @@ def check_annotations(input_vcf: str):
     basename = os.path.splitext(input_vcf)[0]
 
     click.echo("\nPerforming checks... ", nl=False)
-    assert_correct(os.path.join(DATADIR,
-                                "{}_ann_basic.vcf".format(basename)),
-                   os.path.join(DATADIR,
-                                "{}_ann_offline_basic.vcf".format(basename)))
-    assert_correct(os.path.join(DATADIR,
-                                "{}_ann_crossref.vcf".format(basename)),
-                   os.path.join(DATADIR,
-                                "{}_ann_offline_crossref.vcf".format(basename)))
-    assert_correct(os.path.join(DATADIR,
-                                "{}_ann_variab.vcf".format(basename)),
-                   os.path.join(DATADIR,
-                                "{}_ann_offline_variab.vcf".format(basename)))
-    assert_correct(os.path.join(DATADIR,
-                                "{}_ann_predict.vcf".format(basename)),
-                   os.path.join(DATADIR,
-                                "{}_ann_offline_predict.vcf".format(basename)))
-    assert_correct(os.path.join(DATADIR,
-                                "{}_ann.vcf".format(basename)),
-                   os.path.join(DATADIR,
-                                "{}_ann_offline.vcf".format(basename)))
+    assert_correct_vcf(os.path.join(DATADIR, "{}_ann_basic.vcf".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_basic.vcf".format(basename)))
+    assert_correct_csv(os.path.join(DATADIR, "{}_ann_basic.csv".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_basic.csv".format(basename)))
+    assert_correct_vcf(os.path.join(DATADIR, "{}_ann_crossref.vcf".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_crossref.vcf".format(basename)))
+    assert_correct_csv(os.path.join(DATADIR, "{}_ann_crossref.csv".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_crossref.csv".format(basename)))
+    assert_correct_vcf(os.path.join(DATADIR, "{}_ann_variab.vcf".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_variab.vcf".format(basename)))
+    assert_correct_csv(os.path.join(DATADIR, "{}_ann_variab.csv".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_variab.csv".format(basename)))
+    assert_correct_vcf(os.path.join(DATADIR, "{}_ann_predict.vcf".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_predict.vcf".format(basename)))
+    assert_correct_csv(os.path.join(DATADIR, "{}_ann_predict.csv".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline_predict.csv".format(basename)))
+    assert_correct_vcf(os.path.join(DATADIR, "{}_ann.vcf".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline.vcf".format(basename)))
+    assert_correct_csv(os.path.join(DATADIR, "{}_ann.csv".format(basename)),
+                       os.path.join(DATADIR, "{}_ann_offline.csv".format(basename)))
     click.echo("All checks passed.")
 
 
 def launch_annotations(input_vcf: str):
-    """
-    Perform all the annotations (online and offline) on the given VCF file.
+    """Perform all the annotations (online, offline and conversion to
+    CSV) on the given VCF file.
 
     :param str input_vcf: input vcf file name
 
@@ -74,41 +87,36 @@ def launch_annotations(input_vcf: str):
     click.echo("Starting online annotations.")
     annotate(os.path.join(DATADIR, input_vcf),
              os.path.join(DATADIR, "{}_ann_basic.vcf".format(basename)),
-             basic=True)
+             basic=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
              os.path.join(DATADIR, "{}_ann_crossref.vcf".format(basename)),
-             crossref=True)
+             crossref=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
              os.path.join(DATADIR, "{}_ann_variab.vcf".format(basename)),
-             variab=True)
+             variab=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
              os.path.join(DATADIR, "{}_ann_predict.vcf".format(basename)),
-             predict=True)
+             predict=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
-             os.path.join(DATADIR, "{}_ann.vcf".format(basename)))
+             os.path.join(DATADIR, "{}_ann.vcf".format(basename)), csv=True)
     click.echo("Done.")
 
     click.echo("Starting offline annotations.")
     annotate(os.path.join(DATADIR, input_vcf),
-             os.path.join(DATADIR,
-                          "{}_ann_offline_basic.vcf".format(basename)),
-             basic=True, offline=True)
+             os.path.join(DATADIR, "{}_ann_offline_basic.vcf".format(basename)),
+             basic=True, offline=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
-             os.path.join(DATADIR,
-                          "{}_ann_offline_crossref.vcf".format(basename)),
-             crossref=True, offline=True)
+             os.path.join(DATADIR, "{}_ann_offline_crossref.vcf".format(basename)),
+             crossref=True, offline=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
-             os.path.join(DATADIR,
-                          "{}_ann_offline_variab.vcf".format(basename)),
-             variab=True, offline=True)
+             os.path.join(DATADIR, "{}_ann_offline_variab.vcf".format(basename)),
+             variab=True, offline=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
-             os.path.join(DATADIR,
-                          "{}_ann_offline_predict.vcf".format(basename)),
-             predict=True, offline=True)
+             os.path.join(DATADIR, "{}_ann_offline_predict.vcf".format(basename)),
+             predict=True, offline=True, csv=True)
     annotate(os.path.join(DATADIR, input_vcf),
-             os.path.join(DATADIR,
-                          "{}_ann_offline.vcf".format(basename)),
-             offline=True)
+             os.path.join(DATADIR, "{}_ann_offline.vcf".format(basename)),
+             offline=True, csv=True)
     click.echo("Done.")